springboot脚手架（极简框架）
