+++ chapter-1-1/pom.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+
+	<groupId>com.zhbean</groupId>
+	<artifactId>chapter-1-1</artifactId>
+	<version>0.0.1-SNAPSHOT</version>
+	<packaging>jar</packaging>
+
+	<name>chapter-1-1</name>
+	<description>a simple framework for Spring Boot</description>
+
+	<parent>
+		<artifactId>springboot</artifactId>
+		<groupId>com.dxy.zhbean</groupId>
+		<version>0.0.1-SNAPSHOT</version>
+	</parent>
+
+</project>
