+++ chapter-1-1/src/main/java/com/zhbean/demo/DemoApplication.java
@@ -1,0 +1,12 @@
+package com.zhbean.demo;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DemoApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(DemoApplication.class, args);
+	}
+}
