+++ chapter-1-1/src/test/java/com/zhbean/demo/Chapter11ApplicationTests.java
@@ -1,0 +1,16 @@
+package com.zhbean.demo;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class Chapter11ApplicationTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
