+++ chapter-5-1/src/main/java/com/zhbean/jwtauth/model/JwtUser.java
@@ -1,0 +1,17 @@
+package com.zhbean.jwtauth.model;
+
+import lombok.Data;
+
+/**
+ * Created by IntelliJ IDEA.
+ *
+ * @author: zhbean
+ * @Date: 2018/9/13
+ */
+@Data
+public class JwtUser {
+
+    private String id;
+    private String username;
+
+}
