--- jwtauth/src/main/java/com/dxy/zhbean/jwtauth/JwtauthApplication.java
@@ -1,12 +1,0 @@
-package com.dxy.zhbean.jwtauth;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class JwtauthApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(JwtauthApplication.class, args);
-	}
-}
