--- jwtauth/src/main/java/com/dxy/zhbean/jwtauth/model/JwtUser.java
@@ -1,17 +1,0 @@
-package com.dxy.zhbean.jwtauth.model;
-
-import lombok.Data;
-
-/**
- * Created by IntelliJ IDEA.
- *
- * @author: zhbean
- * @Date: 2018/9/13
- */
-@Data
-public class JwtUser {
-
-    private String id;
-    private String username;
-
-}
