--- jwtauth/src/test/java/com/dxy/zhbean/jwtauth/JwtauthApplicationTests.java
@@ -1,16 +1,0 @@
-package com.dxy.zhbean.jwtauth;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class JwtauthApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
