--- springboot-global-exception/src/main/java/com.dxy.zhbean/globalexception/GlobalExceptionApplication.java
@@ -1,12 +1,0 @@
-package com.dxy.zhbean.globalexception;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class GlobalExceptionApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(GlobalExceptionApplication.class, args);
-    }
-}
