--- springboot-global-exception/src/main/java/com.dxy.zhbean/globalexception/http/MultiReadHttpServletRequest.java
@@ -1,100 +1,0 @@
-package com.dxy.zhbean.globalexception.http;
-
-import javax.servlet.ReadListener;
-import javax.servlet.ServletInputStream;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletRequestWrapper;
-import java.io.BufferedReader;
-import java.io.ByteArrayInputStream;
-import java.io.IOException;
-import java.io.InputStreamReader;
-
-/**
- * Created by IntelliJ IDEA.
- *
- * @author: zhbean
- * @Date: 2018/9/5
- */
-public class MultiReadHttpServletRequest extends HttpServletRequestWrapper{
-    private final String body;
-
-    /**
-     * Constructs a request object wrapping the given request.
-     *
-     * @param request The request to wrap
-     * @throws IllegalArgumentException if the request is null
-     */
-    public MultiReadHttpServletRequest(HttpServletRequest request) throws IOException {
-        super(request);
-        //初始化body
-        StringBuilder stringBuilder = new StringBuilder();
-        String characterEncoding = request.getCharacterEncoding();
-        if (characterEncoding == null) {
-            characterEncoding = "UTF-8";
-        }
-        BufferedReader bufferedReader = null;
-        try {
-            ServletInputStream inputStream = request.getInputStream();
-            if (inputStream != null) {
-                bufferedReader = new BufferedReader(new InputStreamReader(inputStream, characterEncoding));
-                char[] chars = new char[128];
-                int readlen = -1;
-                while ((readlen = bufferedReader.read(chars)) > 0) {
-                    stringBuilder.append(chars, 0, readlen);
-                }
-            } else {
-                stringBuilder.append("");
-            }
-        } catch (IOException e) {
-            throw e;
-        } finally {
-            if (bufferedReader != null) {
-                bufferedReader.close();
-            }
-        }
-
-        body = stringBuilder.toString();
-    }
-
-
-    @Override
-    public ServletInputStream getInputStream() throws IOException {
-        final ByteArrayInputStream byteArrayInputStream = new ByteArrayInputStream(body.getBytes());
-        ServletInputStream servletInputStream = new ServletInputStream() {
-            @Override
-            public boolean isFinished() {
-                return false;
-            }
-
-            @Override
-            public boolean isReady() {
-                return false;
-            }
-
-            @Override
-            public void setReadListener(ReadListener listener) {
-
-            }
-
-            @Override
-            public int read() throws IOException {
-                return byteArrayInputStream.read();
-            }
-        };
-        return servletInputStream;
-    }
-
-    @Override
-    public BufferedReader getReader() throws IOException {
-        return new BufferedReader(new InputStreamReader(this.getInputStream()));
-    }
-
-    public String getBody() {
-        return body;
-    }
-
-
-
-
-
-}
