--- springboot-log/src/main/java/com.dxy.zhbean.log/LogApplication.java
@@ -1,12 +1,0 @@
-package com.dxy.zhbean.log;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class LogApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(LogApplication.class, args);
-	}
-}
