--- springboot-log/src/main/java/com.dxy.zhbean.log/aop/CustomsLog.java
@@ -1,18 +1,0 @@
-package com.dxy.zhbean.log.aop;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-/**
- * Created by IntelliJ IDEA.
- *
- * @author: zhbean
- * @Date: 2018/8/28
- */
-@Target(ElementType.METHOD)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface CustomsLog {
-    String value() default "";
-}
