--- springboot-log/src/main/resources/application.yml
@@ -1,8 +1,0 @@
-server:
-  port: 8181
-spring:
-    datasource:
-        url: jdbc:mysql://127.0.0.1:3306/guns?autoReconnect=true&useUnicode=true&characterEncoding=utf8&zeroDateTimeBehavior=CONVERT_TO_NULL&useSSL=false&serverTimezone=UTC
-        username: root
-        password: rootbean
-
